requirements.txt
# Module 2 — Smart Heating Vest Cloud Intelligence
# Python 3.8+

# Data & ML
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Deep learning
tensorflow>=2.10.0

# Firebase
firebase-admin>=6.0.0

# Optional: for .env (Firebase credentials path)
python-dotenv>=1.0.0
